// src/app/models/drinkView.ts
export interface DrinkView {
  titlePageEN: string
  titlePageES: string
  titlePageDE: string
  titlePageFR: string
  titlePageIT: string
  randomBtnEN: string
  randomBtnES: string
  randomBtnDE: string
  randomBtnFR: string
  randomBtnIT: string
  drinksBtnEN: string
  drinksBtnES: string
  drinksBtnDE: string
  drinksBtnFR: string
  drinksBtnIT: string
  loadingEN: string
  loadingES: string
  loadingDE: string
  loadingFR: string
  loadingIT: string
  noDataEN: string
  noDataES: string
  noDataDE: string
  noDataFR: string
  noDataIT: string
}